requirements.txt
transformers==4.37.2
torch==2.1.1
numpy==1.26.4
scikit-build==0.17.6
setuptools==69.1.0
torchvision==0.16.1
addict==2.4.0
lmdb==1.4.1
pyyaml==6.0.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
tensorboard>=2.15.0,<2.17.0
tqdm==4.66.2
psutil==5.9.8
opencv-python-headless==4.9.0.80
basicsr==1.4.2
facexlib==0.3.0
gfpgan==1.3.8
realesrgan==0.3.0
boto3==1.34.40
Pillow==10.2.0
python-dotenv==1.0.1

